control/grpc.py: fix the issue that traddr is not stored in omap
Now, if traddr is not specified in the request, the program will
use the traddr in the conf file, and this traddr is not stored
in OMAP due to the program logic issue. Modify this logic. If
traddr is not specified, assign the traddr in the conf file
to request. In this way, the traddr will be stored in OMAP.
Similarly, store gateway_name to omap.

The discovery controller to be added in the future can obtain
these information from OMAP.

Signed-off-by: Yin Congmin <congmin.yin@intel.com>